Fixing bootstrap JS naming conflict
When bootstrap 2 modal was added in 2.1.1 for future use a GUI tool, there was a conflict with the .modal button class that brings up an iframe modal in certain places in the CP. This has been fixed by giving the modal windows the class of pyro_modal, so you can just switch out the modal class for pyro_modal and everything works the same as in Bootstrap modal.

Fixes Issue #1416.