Making field setup event arguments unnecessary.
When you are creating a field, it either exists or it doesn't (i.e. we have the data for it or we don't). If we don't we don't have any data to pass, but we still need to run events. Field setup events should be aware of the fact that the function probably doesn't have any data to pass.