Add some missing global German translations.
I'm unsure about the $lang['powered_by_pyrocms'] string. It's very common in Germany, so I think a translation is not really necessary. In addition to that, there's no accurate translation for "powered by". The best thing I came up with is "unterstützt von", which translates more to "supported by" than "powered by".

The same with "Keywords", in my opinion it's common enough in Germany to be left unchanged, but that's not my decision to make.