From now on installer translation flags are generated automatically.
If there is a folder such as in the `installer/languages/<lang_name>` and the `<lang_name>` is in the `system/cms/config/language.php` file then there is a translation icon appears in the header of the installer.

The flags icon files are named after the two character code for the representative country of the language. If this behavior needs to be overridden it should be specified in the `$flag_exchange` array on `installer/controllers/installer.php` at line 96.

Signed-off-by: George Petsagourakis <petsagouris@gmail.com>