Added plugin self documentation. Contributions welcome! Search for @todo and you can help document plugins. Follow the structure of the example array and the resulting documentation will be displayed in admin/addons/plugins by clicking on the magnifying glass icon.
Remember to uncomment the array return value!!

For plugins that use dynamic attributes such as {{ foo:bar database_column="where clause" }}
list the database columns that would most likely be used.

For double tags look at the database query that supplies the data and provide
the useful columns that the query selects. (See the blog plugin for an example
of double tag variables).

For plugins that use a __call magic method you can look at Variables or Settings
to get ideas. You may not be able to populate dynamically but may want to
list the common variables instead.