Adding in some alternative parsing options.
The streams parser has a number of options that allow developers to create much more powerful field types. Now the option to parse through the streams parser filter is available in template functions. I had neglected to implement this system-wide as an option until I was reminded while creating documentation.

Also, there is an option to pre_process your view file instead of having the template library build function process it for you.