Refinement to way admin menu is displayed.
Previously, the logic said that if you had a main nav item that had one sub nav, it just displayed that one nav as a main nav item. I don't think that makes a lot of sense, as we can't really know which title we wanted to use. Now it will display a sub menu no matter how many sub menu items there are.