Switching the way that the pages module loads the chunks field type.
The events.php method was just not working, so now we are loading it directly in the pages module, based on if page chunks are being used.

If anyone has a better idea of how to check if page chunks are being used (a setting, maybe?) please let me know.