Font Rendering
Fixed some font rendering and smoothing issues by removing (-webkit-font-smoothing:subpixel-antialiased;) off the <body> and applying it just to the <header>. Chrome still has an issue with font rendering that can not be fixed with out the long awaited patch from Google. However this does fix Safari on desktop and mobile.

Signed-off-by: Joshua Pekera <joshuapekera@me.com>