Sentry is at the point of checking logins and failing to match passwords. This is entirely expected, because Ion Auth was just sha1()'ing a salt and this is instead doing something way more fun with the PHP 5.5 Password Hashing API and it's PHP 5.3 compatible composer package.
I'll write a migration to double-hash all existing users passwords, so if the password fails I can just try it again old-style and see if that works. It'll be slow, but we can replace users passwords over time as they log in with correct passwords, then maybe suggest people send out an email to those remaining old password users in 6 months or whatever.