Expanding the form_output method of the relationship field type to support streams from the pages namespace
The attached code expands the form_output method to take the irregular
table structure of streams in the pages namespace into account when
requesting data for the dropdown menu.
Without this code the relationship fields dropdown lists (used for
pages) show ids as option labels and use the wrong ids to boot.