Fix: return empty when a page is not available in the db
The get method of the page model used to return empty on a missing
page.

Currently it returns an object with a variable called
stream_entry_found, which is added to the page object by the get method
itself.

This causes an issue with the edit method in the pages admin
controller. Quite possibly it causes issues elsewhere too.