Bug Fix: Only set validation rules on submit
The validation rules of the contact plugin were causing issues with
other forms on the same page.

This is easily prevented by setting the validation rules only when the
contact form is actually submitted.

The long list of additions and deletions is due to a change in
indentation, caused by splitting up one if statement into two if
statements on lines 259 and 263.

This simple fix is also included in
https://github.com/pyrocms/pyrocms/pull/2414, but I'd appreciate it if
this could already be part of 2.2.