Fix for page: vars
The issue was that the {{ page:var }} tags were not working because the
page vars were not being loaded until after the page was pulled from
the database. The fix is to push the page vars into the ci cached vars
right after we pull the page entry and before we pull the stream entry.