Adding ability to set a stream in the template library.
When setting a stream in the template library, basically you are
telling Pyro to use the stream parser to process the view. The stream
parser goes in and modifies tags to make things like multiple relations
work (ie, tags with parameters). To see a working exampl, see the blog
module.