Fixing issue with passing rows in streams parse
Passing the row_id as `row_id=id` was causing problems when it was
inside of a variable loop. Changing it to `row_id="{{ id }}"` helps,
however, we now have to deal with variable naming collisions for
overriden stream variables that are not inside tag loops. This is a Lex
issue, however, so it should be fixed in Lex.