Fixing install issue with files module.
The 2nd parameter for integer is $autoIncrement (set to false by
default). Having this set to true means that it is in conflict with the
actual auto_increment primary key.

This seems like it was originally done as a constraint. Maybe this
needs to be a small int or a char - it will only ever be 1 or 0.