Updated reset_pass function to work with Sentry
I updated the reset_pass function to work with Cartalyst Sentry rarer than use the old Ion library. Notice that I send here two emails one with reset code and the other with new password. Before merge this test it before since the process is right but after I change the password using this I can't login and don't know the cause since the password is changed trough Sentry and I try to do all using the library. Maybe this could be improved a lot but is what I think until now