Fix that allows an update to the stream slug
I don't think I have ever figures out why the slug and table name would not update until now. The docs show you can update the slug. I looked at this file and noticed that the $stream parameter was sent to the update method no matter if the data contained a new slug or not. I made this update and everything seems to work now.